Refactor BERT classifier and add loss checker
- Remove custom LyricsPoolingConfig/LyricsPoolingClassifier and
  centralize
  dropout handling via _apply_dropout_overrides
- Adjust BertLikeBuildConfig (remove pooling_strategy, add problem_type)
  and adapt build flow accordingly
- Add src/check_loss_and_weights.py to validate loss wiring and class
  weights
  across tasks
- Fix RandomCropDataCollator randint edge-case and preserve special
  tokens
- Extend test_collator.py for multimodel testing and more diagnostics
- Use float label arrays for multilabel task and print computed class
  weights
  during training runs
- Update many results CSVs (task1..3) and adjust prediction files
  (remove
  task1/task3 preds, tweak task2)
- Minor CLI/output cleanup in bert_based_comparison scripts

+++ src/task2/train_model.py
@@ -40,7 +40,7 @@ set_seed(SEED)
 
 
 def create_label_column(df: pd.DataFrame, label_cols: list[str]) -> pd.Series:
-    return df[label_cols].astype(int).values.tolist()
+    return df[label_cols].astype(float).values.tolist()
 
 
 def main(model_name, baseline=False):
